Enhance Security on recent Created Api

+++ controller/postLikeController.js
@@ -1,14 +1,43 @@
+import blogModel from "../models/blogModel.js";
 import postLikeModel from "../models/postLikeModel.js";
+import userModel from "../models/userModel.js";
 
 export const postLikeSetController = async (req, res) => {
   try {
     // Todo For PostLikeController :
-    // userID and PostID request karni hogi
+    // userID and PostID request karni hogi and also validate it
     // validation check already this user and id present or not
     // if not then save on likepostmodel
     //  set data to frontend developer
 
     const { userId, postId } = req.params;
+
+    if (!postId || !userId) {
+      return res.status(400).send({
+        success: false,
+        message: "All params are required !",
+      });
+    }
+
+    // validate postid and userid on thier model
+
+    const user = await userModel.findById(userId);
+
+    if (!user) {
+      return res.status(400).send({
+        success: false,
+        message: "User doesn't exist ",
+      });
+    }
+
+    const post = await blogModel.findById(postId);
+
+    if (!post) {
+      return res.status(400).send({
+        success: false,
+        message: "post doesn't exist ",
+      });
+    }
 
     const likedCheckForValidation = await postLikeModel.findOne({
       postId: postId,
@@ -48,6 +77,35 @@ export const postLikeDeleteController = async (req, res) => {
     //  set data to frontend developer
 
     const { userId, postId } = req.params;
+
+    // validate postid and userid on thier model
+
+    if (!postId || !userId) {
+      return res.status(400).send({
+        success: false,
+        message: "All params are required !",
+      });
+    }
+
+    // validate postid and userid on thier model
+
+    const user = await userModel.findById(userId);
+
+    if (!user) {
+      return res.status(400).send({
+        success: false,
+        message: "User doesn't exist ",
+      });
+    }
+
+    const post = await blogModel.findById(postId);
+
+    if (!post) {
+      return res.status(400).send({
+        success: false,
+        message: "post doesn't exist ",
+      });
+    }
 
     const unliked = await postLikeModel.findOneAndDelete({
       postId: postId,
@@ -90,6 +148,68 @@ export const postLikeCountController = async (req, res) => {
     res.status(500).send({
       success: false,
       message: "Something Error in Post Like Count Controller",
+    });
+  }
+};
+
+export const postLikeCheckController = async (req, res) => {
+  try {
+    // TODO
+    // 1) Get postid and user id to check and validate it
+    // 2) then check post and user existance in their model
+    // 3) Check user liked or not
+    // 4) then check true and false and return to frontend
+
+    const { postId, userId } = req.params;
+
+    if (!postId || !userId) {
+      return res.status(400).send({
+        success: false,
+        message: "All params are required !",
+      });
+    }
+
+    // validate postid and userid on thier model
+
+    const user = await userModel.findById(userId);
+
+    if (!user) {
+      return res.status(400).send({
+        success: false,
+        message: "User doesn't exist ",
+      });
+    }
+
+    const post = await blogModel.findById(postId);
+
+    if (!post) {
+      return res.status(400).send({
+        success: false,
+        message: "post doesn't exist ",
+      });
+    }
+
+    const postLikedValidate = await postLikeModel.findOne({
+      postId,
+      userId,
+    });
+
+    if (!postLikedValidate) {
+      return res.status(200).send({
+        success: false,
+        message: "user doesn't Like it",
+      });
+    }
+
+    return res.status(200).send({
+      success: true,
+      message: "user liked it",
+    });
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({
+      success: false,
+      message: "Something error in Liked Check Controller",
     });
   }
 };
